+++ src/main/scala/ai/tripl/arc/ARC.scala
@@ -129,8 +129,6 @@ object ARC {
 
     // add spark config to log
     val sparkConf = new java.util.HashMap[String, String]()
-    spark.sparkContext.hadoopConfiguration.set("io.compression.codecs", classOf[ai.tripl.arc.util.ZipCodec].getName)
-
     spark.sparkContext.getConf.getAll.foreach{ case (k, v) => sparkConf.put(k, v) }
 
     implicit val logger = LoggerFactory.getLogger(jobId.getOrElse(spark.sparkContext.applicationId))